refactor(chrome-extension): review feature names
