+++ apps/chrome-extension/src/features/textProfileActiveDropdown/index.ts
@@ -1,0 +1,1 @@
+export { default as TextProfileEditDropdown } from './ui/TextProfileEditDropdown.vue'
