+++ apps/chrome-extension/src/features/textProfileEditDropdown/index.ts
@@ -1,0 +1,1 @@
+export { default as TextProfileEditDropdown } from './ui/TextProfileEditDropdown.vue'
