+++ apps/chrome-extension/src/features/textProfileEditDropdown/ui/TextProfileEditDropdown.vue
