--- apps/chrome-extension/src/features/textProfilesActiveDropdown/index.ts
@@ -1,1 +1,0 @@
-export { default as TextProfilesEditDropdown } from './ui/TextProfilesEditDropdown.vue'
